S10: v_flow modification / deletion
Modification (plan.md §5.2): validates the header SEID via
v_seid_validate on every call (§7 — not just establishment, which
legitimately has no SEID yet), reads+modifies+CAS-writes, retries on
CONFLICT up to V_CAS_MAX_RETRY re-reading fresh state each time, and on
VDP reject/timeout CAS-writes a snapshot of the pre-modification ctx
back so the record ends up correctly reverted (not just abandoned).
That snapshot has nowhere to live in the given struct v_txn, so it's
stashed in t->impl for Stage 1 — documented as a gap needing a real
slot once Stage 2's table claims that field for its own handle.

Deletion (plan.md §5.3): same read-then-CAS-retry shape, but VDP is
called before any Redis mutation, so a VDP reject just fails the
request with nothing to compensate. Accept path frees the TEID (via the
new v_port_pfcp_ctx_teid() accessor — a session's TEID doesn't appear
in Modification/Deletion requests, only in what establishment stored)
and never reclaims the SEID.

Adds v_port_pfcp_ctx_teid() to the port contract with the same
documented-adaptation treatment as the other additions.

test/test_flow_mod_del.c covers accept, VDP-reject compensation
(verifies both the version bump AND that the data byte-for-byte matches
pre-modification), not-found, the free5gc-class SEID fuzz case, deletion
accept/TEID-free and VDP-reject, plus the two plan.md §9.3 scenarios:
a concurrent-modification storm verified via the version counter (can
only reach +2 if neither write was lost) and modification-racing-
deletion (hard invariant: the session always ends up deleted).

+++ test/test_dispatch.c
@@ -81,7 +81,7 @@ void test_dispatch_any_worker(void)
     CHECK((int)(p1 + p2 + p3) == N);
 
     unsigned sent = v_dispatch_io_drain_tx(100);
-    CHECK(sent == N);
+    CHECK((int)sent == N);
 
     printf("test_dispatch_any_worker: done\n");
 }
